refactor: extract dto
