+++ src/model/mod.rs
@@ -1,2 +1,1 @@
-pub mod login_form;
 pub mod user;
